pyproject.toml
[project]
name = "artisan-market"
version = "2025.0"
description = "Polyglot persistence project for handmade goods marketplace"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "pymongo>=4.6.1",
    "redis>=5.0.1",
    "neo4j>=5.16.0",
    "pandas>=2.1.4",
    "numpy>=1.26.3",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.25",
    "sentence-transformers>=2.2.2",
    "pgvector>=0.2.4",
    "click>=8.1.7",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "ruff>=0.1.9",
    "ipython>=8.19.0",
    "pytest-asyncio>=0.21.1",
]

[project.scripts]
load-data = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]